Meta's Shepherd
Meta's Shepherd pretending t have this back in 2023, doesn't add up. If they did its functions are way off and are only a prompt recalibration system which is actually one small part of SSPS but it only provided the prompts recalibration for the user. Still its an attempt to sow evolution. It fails. Happens again I will be sure to up the damages.